backend/requirements.txt
django-annoying==0.10.4
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
